requirements.txt
Flask-JWT-Extended
Flask-RESTful
Flask-SQLAlchemy
Flask-Cors
python-decouple
requests